DCF Valuation: Persistent Systems Ltd.
Persistent Systems Ltd. was valued using a Discounted Cash Flow (DCF) model, employing a WACC of 12.5%. Key assumptions included a Risk-Free Rate of 7% and a Beta of 1.21. Projections for EBIT(1-T), Depreciation, and Capital Expenditure were used to derive Free Cash Flow to Firm (FCFF) through March 2030. The DCF analysis yielded a present value of cash flows of 15,893 and a terminal value of 160,193, with a present value of terminal value at 79,018. This led to an Enterprise Value of 94,911 and an Equity Value of 95,625. The intrinsic value per share was calculated at 5,805.54. Comparing this to the current market price of 5,700 , and considering a projected return of 1.85% , the recommendation for Persistent Systems Ltd.